requirements.txt
paho-mqtt==2.1.0
requests==2.32.3
flask==3.1.0
pyopenssl==24.3.0
pycurl==7.45.6
gunicorn==23.0.0
pytest==8.3.4
pytest-asyncio==1.3.0
python-dotenv==1.0.1
